requirements.txt

numpy
torch
torchvision
sklearn
pandas
seaborn
Pillow
matplotlib
ptitprince
git+https://github.com/sebastianberns/cleanfeatures
git+https://github.com/openai/CLIP.git
dreamsim
